Class naming unification in GraphQL Types (#7, 2 and 5)
- Rename files types/X_type.rb to types/X.rb
- Rename classes Spree::GraphQL::XType to Spree::GraphQL::Types::X
- Add module Spree::GraphQL[::Types] where it was missing
