+++ lib/solidus_graphql_api/graphql/schema.rb
@@ -1,8 +1,8 @@
 module Spree
   module GraphQL
     class Schema < ::GraphQL::Schema
-      mutation(::Spree::GraphQL::MutationType)
-      query(::Spree::GraphQL::QueryType)
+      mutation(::Spree::GraphQL::Types::Mutation)
+      query(::Spree::GraphQL::Types::Query)
     end
   end
 end
