+++ lib/solidus_graphql_api/graphql/types/attachment.rb
@@ -1,0 +1,16 @@
+module Spree
+  module GraphQL
+    module Types
+      class Attachment < ::GraphQL::Schema::Object
+        graphql_name 'Attachment'
+
+        field :id, ID, null: false
+        field :url, String, null: false
+        field :default_url, String, null: false
+        field :default_style,String, null: false
+        field :path, String, null: false
+
+      end
+    end
+  end
+end
