--- lib/solidus_graphql_api/graphql/types/attachment_type.rb
@@ -1,15 +1,0 @@
-module Spree
-  module GraphQL
-    class Attachment < ::GraphQL::Schema::Object
-    graphql_name "Attachment"
-
-      field :id, ID, null: false
-      field :url, String, null: false
-      field :default_url, String, null: false
-      field :default_style,String, null: false
-      field :path, String, null: false
-
-    end
-  end
-end
-
