+++ lib/solidus_graphql_api/graphql/types/base_input_object.rb
@@ -1,2 +1,8 @@
-class Types::BaseInputObject < ::GraphQL::Schema::InputObject
+module Spree
+  module GraphQL
+    module Types
+      class BaseInputObject < ::GraphQL::Schema::InputObject
+      end
+    end
+  end
 end
