+++ lib/solidus_graphql_api/graphql/types/base_interface.rb
@@ -1,3 +1,9 @@
-module Types::BaseInterface
-  include ::GraphQL::Schema::Interface
+module Spree
+  module GraphQL
+    module Types
+      module BaseInterface
+        include ::GraphQL::Schema::Interface
+      end
+    end
+  end
 end
