+++ lib/solidus_graphql_api/graphql/types/base_object.rb
@@ -1,5 +1,9 @@
-module Types
-  class BaseObject < ::GraphQL::Schema::Object
-    global_id_field :id
+module Spree
+  module GraphQL
+    module Types
+      class BaseObject < ::GraphQL::Schema::Object
+        global_id_field :id
+      end
+    end
   end
 end
