+++ lib/solidus_graphql_api/graphql/types/base_union.rb
@@ -1,2 +1,8 @@
-class Types::BaseUnion < ::GraphQL::Schema::Union
+module Spree
+  module GraphQL
+    module Types
+      class BaseUnion < ::GraphQL::Schema::Union
+      end
+    end
+  end
 end
