+++ lib/solidus_graphql_api/graphql/types/image.rb
@@ -1,0 +1,11 @@
+module Spree
+  module GraphQL
+    module Types
+      require_relative './attachment'
+      class Image < BaseObject
+        graphql_name 'Image'
+        field :attachment, Attachment, null: true
+      end
+    end
+  end
+end
