--- lib/solidus_graphql_api/graphql/types/image_type.rb
@@ -1,9 +1,0 @@
-module Spree
-  module GraphQL
-    require_relative './attachment_type'
-    class ImageType < ::Types::BaseObject
-      graphql_name "Image"
-      field :attachment, Attachment, null: true
-    end
-  end
-end
