--- lib/solidus_graphql_api/graphql/types/mutation_type.rb
@@ -1,9 +1,0 @@
-module Spree
-  module GraphQL
-    class MutationType < ::Types::BaseObject
-      field :create_order, mutation: ::Spree::GraphQL::Mutations::Order::CreateOrder
-      field :add_or_update_item, mutation: ::Spree::GraphQL::Mutations::Order::AddOrUpdateCartItem
-      field :remove_item, mutation: ::Spree::GraphQL::Mutations::Order::RemoveItemFromCart
-    end
-  end
-end
