+++ lib/solidus_graphql_api/graphql/types/option.rb
@@ -1,0 +1,14 @@
+module Spree
+  module GraphQL
+    module Types
+      class Option < BaseObject
+        graphql_name 'Option'
+
+        field :id, ID, null: false
+        field :graphql_name, String, null: false
+        field :presentation, String, null: true
+        field :position, Integer, null: true
+      end
+    end
+  end
+end
