--- lib/solidus_graphql_api/graphql/types/option_type.rb
@@ -1,12 +1,0 @@
-module Spree
-  module GraphQL
-    class OptionType < ::Types::BaseObject
-      graphql_name "OptionType"
-
-      field :id, ID, null: false
-      field :graphql_name, String, null: false
-      field :presentation, String, null: true
-      field :position, Integer, null: true
-    end
-  end
-end
