+++ lib/solidus_graphql_api/graphql/types/shipment_method.rb
@@ -1,8 +1,10 @@
 module Spree
   module GraphQL
-    class ShipmentMethodType < ::Types::BaseObject
-       field :id, ID, null: false
-       field  :name, String, null: false
+    module Types
+      class ShipmentMethod < BaseObject
+        field :id, ID, null: false
+        field :name, String, null: false
+      end
     end
   end
 end
