--- lib/solidus_graphql_api/graphql/types/style_type.rb
@@ -1,10 +1,0 @@
-module Spree
-  module GraphQL
-    class Style < ::GraphQL::Schema::Object
-      field :style_graphql_name, String, null: true
-      field :geometry, String, null: true
-      field :url, String, null: false
-    end
-  end
-end
-
